% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_num.R
\name{f_num}
\alias{f_num}
\title{Format number with abreviation and separators}
\usage{
f_num(
  x,
  big = ",",
  dec = ".",
  thousand = "K",
  million = "M",
  billion = "B",
  dig = 0
)
}
\arguments{
\item{x}{Object with values}

\item{big}{Thousand separator}

\item{dec}{Decimal separator}

\item{thousand}{Abreviation for thousands}

\item{million}{Abreviation for millions}

\item{billion}{Abreviation for billions}

\item{dig}{Digits after decimal mark}
}
\description{
Return is character class
}
\examples{
f_num(12345678956, billion = 'G')
f_num(512347896, million = 'Mi')
f_num(9995198, thousand = 'm', dig = 3, dec = ',', big = '.')
f_num(55566312345678956, billion = 'G')
f_num(Inf)
f_num(-Inf)

}
