% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fina.R
\name{fina}
\alias{fina}
\title{First n value taking NA's into account}
\usage{
fina(x, n = 1L, na_rm = TRUE)
}
\arguments{
\item{x}{Objeto com valores}

\item{n}{Position to return, default is 1 (first). If zero or less it will be corrected to 1}

\item{na_rm}{If TRUE or T will remove NA's. Default value is T.}
}
\description{
If all values are NA returns NA
}
\examples{
fina(c(5, 1, 2, 3, NA))
fina(c(5, 1, 2, 3, NA), n = 2)
fina(c(NA, 1, 2, 3, NA), na_rm = TRUE)
fina(c(NA, 1, 2, 3, NA), na_rm = FALSE)
fina(c(NA, NA, NA), na_rm = TRUE)
fina(c(NA, NA, NA), na_rm = FALSE)

}
