% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_info.R
\name{gt_info}
\alias{gt_info}
\title{Show in interactive gt table a data.frame from df_info function}
\usage{
gt_info(df)
}
\arguments{
\item{df}{A data.frame object from df_info function}
}
\description{
Show in interactive gt table a data.frame from df_info function
}
\examples{
mtcars |> df_info() |> gt_info()

}
