% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_name.R
\name{is_valid_name}
\alias{is_valid_name}
\title{Test if values are valid names}
\usage{
is_valid_name(x)
}
\arguments{
\item{x}{values to test}
}
\description{
Test if values are valid names
}
\examples{
'abc ' |> is_valid_name()
'abc' |> is_valid_name()
c(1, 'a', 'b') |> is_valid_name()
list('1', 2, 3, NA, 'a') |> is_valid_name()

}
