% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lana.R
\name{lana}
\alias{lana}
\title{Last n value taking NA's into account}
\usage{
lana(x, n = 1, na_rm = TRUE)
}
\arguments{
\item{x}{Object with values}

\item{n}{Position to return, default is 1 (last). If zero or less it will be corrected to 1.}

\item{na_rm}{If TRUE or T removes the NA's. Default value is T.}
}
\description{
If all values are NA returns NA
}
\examples{
lana(c(5, 1, 2, 3, NA))
lana(c(NA, 1, 2, 3, NA), na_rm = FALSE)
lana(c(5, 1, 2, 3, NA), n = 3)
lana(c(5, 1, 2, 3, NA), n = 3)
lana(c(NA, NA, NA), na_rm = TRUE)
lana(c(NA, NA, NA), na_rm = FALSE)

}
