% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mana.R
\name{mana}
\alias{mana}
\title{Maximum value taking NA's into account}
\usage{
mana(x, na_rm = TRUE)
}
\arguments{
\item{x}{Object with values}

\item{na_rm}{If TRUE or T removes NA's. Default value is T.}
}
\description{
Maximum value taking NA's into account
}
\examples{
mana(c(-1, 2, 3, NA))
mana(c(1, 2, 3, NA), na_rm = FALSE)
mana(c(1, 2, 3, NA, NaN), na_rm = FALSE)

}
