% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mina.R
\name{mina}
\alias{mina}
\title{Minimum value taking NA's into account}
\usage{
mina(x, na_rm = TRUE)
}
\arguments{
\item{x}{Object with values}

\item{na_rm}{If TRUE or T removes NA's. Default value is T.}
}
\description{
Minimum value taking NA's into account
}
\examples{
mina(c(-1, 2, 3, NA))
mina(c(1, 2, 3, NA), na_rm = FALSE)
mina(c(1, 2, 3, NA, NaN), na_rm = FALSE)

}
