% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{pn}
\alias{pn}
\alias{p10}
\alias{p25}
\alias{p75}
\alias{p90}
\alias{p95}
\alias{p99}
\title{Functions to calculate percentiles}
\usage{
pn(y, p, na_rm = T, ...)

p10(y, na_rm = T, ...)

p25(y, na_rm = T, ...)

p75(y, na_rm = T, ...)

p90(y, na_rm = T, ...)

p95(y, na_rm = T, ...)

p99(y, na_rm = T, ...)
}
\arguments{
\item{y}{Object with values}

\item{p}{Probs as in quantile function}

\item{na_rm}{Removes NAs}

\item{...}{as in quantile function}
}
\description{
Functions to calculate percentiles
}
\examples{

p25(1:3)
p75(1:4)
p99(0:100)

}
