% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spada.R
\name{spada}
\alias{spada}
\title{Launch a 'shiny' application for data analysis}
\usage{
spada(..., run_local = TRUE)
}
\arguments{
\item{...}{Objects of data.frame class}

\item{run_local}{Logical. Whether to run the application locally. If TRUE the app will close at session end.}
}
\value{
An object of class 'shiny.appobj' representing the 'shiny' application.
  Printing the object launches the interactive app in a web browser.
}
\description{
Generates a 'shiny' application for interactive data analysis.
}
\examples{
if(interactive()) spada(datasets::mtcars)

}
