% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suna.R
\name{suna}
\alias{suna}
\title{Sum taking NA's into account}
\usage{
suna(x, na_rm = TRUE)
}
\arguments{
\item{x}{Object to be summed}

\item{na_rm}{If TRUE or T removes NA's. Default value is T.}
}
\description{
Sum taking NA's into account
}
\examples{
suna(c(1, 2, 3, NA))
suna(c(1, 2, 3, NA), na_rm = FALSE)

}
