% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spada.R
\name{spada}
\alias{spada}
\title{Spada (Data Analysis)}
\usage{
spada(...)
}
\arguments{
\item{...}{Objects of data.frame class}
}
\description{
Function that generates a Shiny App for Data Analysis
}
\examples{
if(interactive()) spada(datasets::mtcars)

}
